feat: some code style fixes
