+++ tools/src/main/java/ru/mintrocket/lib/mintpermissions/tools/uirequests/internal/ViewModeEnabledObserver.kt
@@ -1,0 +1,19 @@
+package ru.mintrocket.lib.mintpermissions.tools.uirequests.internal
+
+import androidx.lifecycle.DefaultLifecycleObserver
+import androidx.lifecycle.LifecycleOwner
+
+internal class ViewModeEnabledObserver(
+    private val viewModel: UiRequestViewModel<*, *>
+) : DefaultLifecycleObserver {
+
+    override fun onResume(owner: LifecycleOwner) {
+        super.onResume(owner)
+        viewModel.setEnabled(true)
+    }
+
+    override fun onPause(owner: LifecycleOwner) {
+        super.onPause(owner)
+        viewModel.setEnabled(false)
+    }
+}
